+++ Utility.ino
@@ -32,7 +32,7 @@ bool exists_in_byte_array(uint8_t value, const uint8_t arr[], uint8_t arr_len,
       }  // algorithm pooped out or value doesn't exist.
     }
   } else {  // raw search
-    for (int i = 0; i < arr_len; i++) {
+    for (uint8_t i = 0; i < arr_len; i++) {
       if (arr[i] == value) {
         return true;
       }
